+++ src/FileInterface.cpp
@@ -10,7 +10,7 @@
 #include <fstream>
 #include "graylog_logger/FileInterface.hpp"
 
-FileInterface::FileInterface(std::string fileName, int maxQueueLength) : BaseLogHandler(maxQueueLength), fileName(fileName), fileThread(&FileInterface::ThreadFunction, this) {
+FileInterface::FileInterface(std::string fileName, size_t maxQueueLength) : BaseLogHandler(maxQueueLength), fileName(fileName), fileThread(&FileInterface::ThreadFunction, this) {
     
 }
